fix: Catch Display Moved when looking for orphaned spaces.
When adding or removing displays (docking and undocking) MacOS sends a
series of events to notify about it. For example:
  Display 1 Configured (internal display)
  Display 2 Removed (external display)
  Display 1 Added (inherits the existing spaces from 2)
Here Paneru catches the event and remaps the existing windows to new
display.

Sometimes this sequence is different:
  Display 1 Configured (internal display)
  Display 2 Removed (external display)
  Display 1 Added (has new empty workspace!)
  # Paneru can't find anything.
  Display 1 Moved (here it finally gets the workspaces)
  
This fix should also catch movement of displays and correctly update
internal state.