fix: Catch dragged window marker to move windows between displays.
No longer is it the last focused window which is thrown to another
display - which may result in stray relocations.
The current mechanism is: while draging a window, there is a short
lived marker (on a Timer) which specifies that a window is receiving a
drag event and that this window can be moved to another display during a
display change event.